refactor: move code from .__init__ files and remove unnecessary nesting in tests/ directories
