--- src/authentication/__init__.py
@@ -1,3 +1,0 @@
-ANAFIC = "anafic"
-ADL = "adl"
-MAJOR = "major"
