+++ src/authentication/constants.py
@@ -1,0 +1,3 @@
+ANAFIC = "anafic"
+ADL = "adl"
+MAJOR = "major"
