+++ src/authentication/tests/integration_tests/test_adl_active_api_view.py
@@ -1,9 +1,9 @@
 from parameterized import parameterized
 from rest_framework.reverse import reverse
 
-from authentication import ADL, MAJOR
-from authentication.tests import CouchdbUserFactory
-from grm.tests import BaseTestCase
+from authentication.constants import ADL, MAJOR
+from authentication.factories import CouchdbUserFactory
+from grm.tests.base import BaseTestCase
 
 
 class TestADLActiveAPIView(BaseTestCase):
