+++ src/dashboard/participatory_budget/tests/integration_test/test_dashboard_template_view.py
